+++ feature/onboarding/src/main/java/com/napzak/market/onboarding/genre/GenreScreen.kt
@@ -53,8 +53,8 @@ import com.napzak.market.onboarding.genre.component.GenreGridList
 import com.napzak.market.onboarding.genre.model.GenreEvent
 import com.napzak.market.onboarding.genre.model.GenreUiModel
 import com.napzak.market.onboarding.genre.model.GenreUiState
-import com.napzak.market.util.android.model.ShadowDirection
-import com.napzak.market.util.android.napzakGradientShadow
+import com.napzak.market.ui_util.ShadowDirection
+import com.napzak.market.ui_util.napzakGradientShadow
 import com.napzak.market.util.android.noRippleClickable
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
